add lazy storage for SqlStatements
