--- src/main/java/org/fisheep/bean/data/Structs.java
@@ -1,20 +1,0 @@
-package org.fisheep.bean.data;
-
-import org.eclipse.serializer.reference.Lazy;
-import org.fisheep.bean.SqlStatement;
-import org.fisheep.bean.Struct;
-
-import java.util.HashMap;
-import java.util.List;
-
-/**
- * @author BigOrange
- */
-public class Structs {
-
-    private Lazy<HashMap<SqlStatement, List<Struct>>> map;
-
-    private void add() {
-
-    }
-}
