+++ src/main/java/org/fisheep/logic/ExplainManager.java
@@ -1,0 +1,44 @@
+package org.fisheep.logic;
+
+import io.javalin.http.Context;
+import org.fisheep.bean.SqlStatement;
+import org.fisheep.common.SealException;
+import org.fisheep.common.StorageManagerFactory;
+import org.fisheep.util.PcapUtil;
+
+import java.util.ArrayList;
+import java.util.List;
+import java.util.concurrent.CompletableFuture;
+
+public class ExplainManager {
+
+    //todo
+    public static void upload(Context ctx) throws SealException {
+        var uploadedFile = ctx.uploadedFile("file");
+        var id = Integer.parseInt(ctx.formParam("id"));
+        var data = StorageManagerFactory.data();
+        String timestamp = data.dbs().addTimestamp(id);
+        var db = data.dbs().one(id);
+        //再测试下连接
+        DbManager.getDbVersion(db);
+        data.sqlStatements().add(db.getId() + timestamp, new ArrayList<>());
+        ctx.async(() -> {
+            ctx.result("Task received and is being processed");
+            // 使用CompletableFuture异步处理任务
+            CompletableFuture.runAsync(() -> {
+                try {
+                    var results = PcapUtil.parseLogFile(uploadedFile, db.getPort());
+                    data.sqlStatements().add(db.getId() + timestamp, results);
+                } catch (SealException e) {
+                    //todo 加一个状态map
+                    throw new RuntimeException(e);
+                }
+            });
+            List<SqlStatement> all = data.sqlStatements().all(db.getId() + timestamp);
+            all.forEach(sqlStatement ->
+                    CompletableFuture.runAsync(() -> {
+                        //todo 解析SQL
+                    }));
+        });
+    }
+}
